use newsletter api calls
